before copying file, we check now if the files is already in the working dir

+++ src/p8microclimate_gui.cpp
@@ -80,15 +80,21 @@ void p8microclimate_gui::on_pb_map_released()
     QFileInfo finfo = QFileInfo(fname);
     ui->le_map->setText(finfo.fileName());
     this->p8microclimate->setParameterValue("MapPic",finfo.fileName().toStdString());
-    if( QFile::exists(settings.value("workPath").toString() + finfo.fileName()))
+    QFileInfo workfinfo = QFileInfo(QString(settings.value("workPath").toString() + "/" + finfo.fileName()));
+    if(finfo.absolutePath() != workfinfo.absolutePath())
     {
-        QFile::remove(settings.value("workPath").toString() +"/"+ finfo.fileName());
-        QFile::copy(fname,settings.value("workPath").toString() +"/"+ finfo.fileName());
+        if( QFile::exists(settings.value("workPath").toString() + finfo.fileName()))
+        {
+            QFile::remove(settings.value("workPath").toString() +"/"+ finfo.fileName());
+            QFile::copy(fname,settings.value("workPath").toString() +"/"+ finfo.fileName());
+        }
+        else
+        {
+            QFile::copy(fname,settings.value("workPath").toString() +"/"+ finfo.fileName());
+        }
     }
-    else
-    {
-        QFile::copy(fname,settings.value("workPath").toString() +"/"+ finfo.fileName());
-    }
+    settings.setValue("dataPath",finfo.absolutePath());
+
 }
 
 void p8microclimate_gui::on_pb_shape_released()
@@ -96,21 +102,27 @@ void p8microclimate_gui::on_pb_shape_released()
     QSettings settings;
     this->p8microclimate->workingDir = settings.value("workPath").toString().toStdString();
     QString datapath = settings.value("dataPath").toString() + "/";
-    QString fname = QFileDialog::getOpenFileName(this,"Shape File", QString(this->p8microclimate->workingDir.c_str()), "*.sh");
+    QString fname = QFileDialog::getOpenFileName(this,"Shape File", datapath, "*.sh");
     if (fname == "")
         return;
     QFileInfo finfo = QFileInfo(fname);
     ui->le_shape->setText(finfo.fileName());
     this->p8microclimate->setParameterValue("Shapefile",finfo.fileName().toStdString());
-    if( QFile::exists(settings.value("workPath").toString() + finfo.fileName()))
+    QFileInfo workfinfo = QFileInfo(QString(settings.value("workPath").toString() + "/" + finfo.fileName()));
+    if(finfo.absolutePath() != workfinfo.absolutePath())
     {
-        QFile::remove(settings.value("workPath").toString() +"/"+ finfo.fileName());
-        QFile::copy(fname,settings.value("workPath").toString() +"/"+ finfo.fileName());
+        if( QFile::exists(settings.value("workPath").toString() + finfo.fileName()))
+        {
+            QFile::remove(settings.value("workPath").toString() +"/"+ finfo.fileName());
+            QFile::copy(fname,settings.value("workPath").toString() +"/"+ finfo.fileName());
+        }
+        else
+        {
+            QFile::copy(fname,settings.value("workPath").toString() +"/"+ finfo.fileName());
+        }
     }
-    else
-    {
-        QFile::copy(fname,settings.value("workPath").toString() +"/"+ finfo.fileName());
-    }
+    settings.setValue("dataPath",finfo.absolutePath());
+
 }
 
 void p8microclimate_gui::on_pb_landuse_released()
@@ -118,21 +130,26 @@ void p8microclimate_gui::on_pb_landuse_released()
     QSettings settings;
     this->p8microclimate->workingDir = settings.value("workPath").toString().toStdString();
     QString datapath = settings.value("dataPath").toString() + "/";
-    QString fname = QFileDialog::getOpenFileName(this,"Landuse File",QString(this->p8microclimate->workingDir.c_str()),"*.txt");
+    QString fname = QFileDialog::getOpenFileName(this,"Landuse File",datapath,"*.txt");
     if(fname == "")
         return;
     QFileInfo finfo = QFileInfo(fname);
     ui->le_landuse->setText(finfo.fileName());
     this->p8microclimate->setParameterValue("Landuse",finfo.fileName().toStdString());
-    if( QFile::exists(settings.value("workPath").toString() + finfo.fileName()))
+    QFileInfo workfinfo = QFileInfo(QString(settings.value("workPath").toString() + "/" + finfo.fileName()));
+    if(finfo.absolutePath() != workfinfo.absolutePath())
     {
-        QFile::remove(settings.value("workPath").toString() +"/"+ finfo.fileName());
-        QFile::copy(fname,settings.value("workPath").toString() +"/"+ finfo.fileName());
+        if( QFile::exists(settings.value("workPath").toString() + finfo.fileName()))
+        {
+            QFile::remove(settings.value("workPath").toString() +"/"+ finfo.fileName());
+            QFile::copy(fname,settings.value("workPath").toString() +"/"+ finfo.fileName());
+        }
+        else
+        {
+            QFile::copy(fname,settings.value("workPath").toString() +"/"+ finfo.fileName());
+        }
     }
-    else
-    {
-        QFile::copy(fname,settings.value("workPath").toString() +"/"+ finfo.fileName());
-    }
+    settings.setValue("dataPath",finfo.absolutePath());
 }
 
 void p8microclimate_gui::on_pb_wsud_released()
